+++ ui/src/pages/DataLogs/ShareQueryResultPage.tsx
@@ -1,0 +1,8 @@
+import QueryResult from "@/pages/DataLogs/components/QueryResult";
+import useLogUrlParams from "@/pages/DataLogs/hooks/useLogUrlParams";
+
+const ShareQueryResultPage = () => {
+  useLogUrlParams();
+  return <QueryResult />;
+};
+export default ShareQueryResultPage;
